WIP quiver client and quiver router
I realized that the primary use case in a React app is significantly
different than the primary use case in Node. In Node we want something
that's very, very simple, all-in-one. In a React app we want to be able
to register the clients and routers separately from the functions so
that we can have a kind of "micro app" situation.